# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sparkle Note Hub API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Note {
  id: string;
  title: string;
  content: string;
  /** Hex color or named color for the note card */
  color: string;
  tags: string[];
  pinned: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface CreateNoteBody {
  title: string;
  content: string;
  color?: string;
  tags?: string[];
  pinned?: boolean;
}

export interface UpdateNoteBody {
  title?: string;
  content?: string;
  color?: string;
  tags?: string[];
  pinned?: boolean;
}

export interface NoteStats {
  total: number;
  pinned: number;
  withTags: number;
  /** Notes modified in last 7 days */
  recentCount: number;
}

export interface TagCount {
  tag: string;
  count: number;
}

export interface ErrorResponse {
  error: string;
}

export type ListNotesParams = {
  /**
   * Full-text search across title and content
   */
  search?: string;
  /**
   * Filter notes by tag
   */
  tag?: string;
  /**
   * Filter only pinned notes
   */
  pinned?: boolean;
};