# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sparkle Note Hub API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Returns all notes, optionally filtered by search or tag
 * @summary List all notes
 */
export const ListNotesQueryParams = zod.object({
  search: zod.coerce
    .string()
    .optional()
    .describe("Full-text search across title and content"),
  tag: zod.coerce.string().optional().describe("Filter notes by tag"),
  pinned: zod.coerce.boolean().optional().describe("Filter only pinned notes"),
});

export const ListNotesResponseItem = zod.object({
  id: zod.string(),
  title: zod.string(),
  content: zod.string(),
  color: zod.string().describe("Hex color or named color for the note card"),
  tags: zod.array(zod.string()),
  pinned: zod.boolean(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const ListNotesResponse = zod.array(ListNotesResponseItem);

/**
 * @summary Create a note
 */
export const CreateNoteBody = zod.object({
  title: zod.string(),
  content: zod.string(),
  color: zod.string().optional(),
  tags: zod.array(zod.string()).optional(),
  pinned: zod.boolean().optional(),
});

/**
 * Returns counts and aggregations for dashboard
 * @summary Get note statistics
 */
export const GetNoteStatsResponse = zod.object({
  total: zod.number(),
  pinned: zod.number(),
  withTags: zod.number(),
  recentCount: zod.number().describe("Notes modified in last 7 days"),
});

/**
 * Returns all distinct tags used across notes
 * @summary Get all unique tags
 */
export const GetAllTagsResponseItem = zod.object({
  tag: zod.string(),
  count: zod.number(),
});
export const GetAllTagsResponse = zod.array(GetAllTagsResponseItem);

/**
 * Returns the 5 most recently updated notes
 * @summary Get recently modified notes
 */
export const GetRecentNotesResponseItem = zod.object({
  id: zod.string(),
  title: zod.string(),
  content: zod.string(),
  color: zod.string().describe("Hex color or named color for the note card"),
  tags: zod.array(zod.string()),
  pinned: zod.boolean(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const GetRecentNotesResponse = zod.array(GetRecentNotesResponseItem);

/**
 * @summary Get a single note
 */
export const GetNoteParams = zod.object({
  id: zod.coerce.string(),
});

export const GetNoteResponse = zod.object({
  id: zod.string(),
  title: zod.string(),
  content: zod.string(),
  color: zod.string().describe("Hex color or named color for the note card"),
  tags: zod.array(zod.string()),
  pinned: zod.boolean(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Update a note
 */
export const UpdateNoteParams = zod.object({
  id: zod.coerce.string(),
});

export const UpdateNoteBody = zod.object({
  title: zod.string().optional(),
  content: zod.string().optional(),
  color: zod.string().optional(),
  tags: zod.array(zod.string()).optional(),
  pinned: zod.boolean().optional(),
});

export const UpdateNoteResponse = zod.object({
  id: zod.string(),
  title: zod.string(),
  content: zod.string(),
  color: zod.string().describe("Hex color or named color for the note card"),
  tags: zod.array(zod.string()),
  pinned: zod.boolean(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Delete a note
 */
export const DeleteNoteParams = zod.object({
  id: zod.coerce.string(),
});

/**
 * @summary Toggle note pin status
 */
export const ToggleNotePinParams = zod.object({
  id: zod.coerce.string(),
});

export const ToggleNotePinResponse = zod.object({
  id: zod.string(),
  title: zod.string(),
  content: zod.string(),
  color: zod.string().describe("Hex color or named color for the note card"),
  tags: zod.array(zod.string()),
  pinned: zod.boolean(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});