# lib/api-zod/src/generated/types/createNoteBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sparkle Note Hub API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateNoteBody {
  title: string;
  content: string;
  color?: string;
  tags?: string[];
  pinned?: boolean;
}