# lib/api-zod/src/generated/types/note.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sparkle Note Hub API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Note {
  id: string;
  title: string;
  content: string;
  /** Hex color or named color for the note card */
  color: string;
  tags: string[];
  pinned: boolean;
  createdAt: Date;
  updatedAt: Date;
}