# lib/api-zod/src/generated/types/noteStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sparkle Note Hub API specification
 * OpenAPI spec version: 0.1.0
 */

export interface NoteStats {
  total: number;
  pinned: number;
  withTags: number;
  /** Notes modified in last 7 days */
  recentCount: number;
}