# lib/api-zod/src/generated/types/updateNoteBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sparkle Note Hub API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateNoteBody {
  title?: string;
  content?: string;
  color?: string;
  tags?: string[];
  pinned?: boolean;
}